auto add kubernetes version
